Move to jest and npm
