+++ package.json
@@ -4,10 +4,8 @@
   "description": "configurationLoader loads configurations from a json file and then merges them with existing configurations",
   "main": "dist/configurationLoader.js",
   "scripts": {
-    "webpack": "node ./node_modules/webpack/bin/webpack",
-    "build": "npm run webpack",
-    "prepublishOnly": "npm install --package-lock-only && yarn install --package-lock-only && npm run build",
-    "test": "node node_modules/.bin/mocha --opts test/mocha.opts"
+    "build": "rollup -c",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -24,17 +22,12 @@
   },
   "homepage": "https://github.com/biotope/biotope-configuration",
   "devDependencies": {
-    "@types/chai": "^4.1.2",
-    "@types/mocha": "^2.2.48",
-    "awesome-typescript-loader": "^4.0.0",
-    "chai": "^4.1.2",
+    "@types/jest": "^23.3.12",
     "deepmerge": "^2.1.1",
-    "mocha": "^5.0.1",
-    "ts-node": "^6.1.0",
-    "typescript": "^2.9.1",
-    "webpack": "4.12.0",
-    "webpack-cli": "^3.0.3",
-    "whatwg-fetch": "^2.0.4"
+    "jest": "^23.6.0",
+    "rollup": "^1.1.0",
+    "ts-jest": "^23.10.5",
+    "typescript": "^2.9.2"
   },
   "dependencies": {}
 }
