+++ src/deepGet.spec.ts
@@ -1,12 +1,11 @@
-import {expect} from 'chai';
-import {deepGet} from './deepGet';
+import { deepGet } from './deepGet';
 
 describe('#deepGet', () => {
   it('returns initial object for no key', () => {
     const initialObject = {};
     const returnedValue = deepGet(initialObject, undefined);
 
-    expect(returnedValue).to.deep.equal(initialObject);
+    expect(returnedValue).toEqual(initialObject);
   });
 
   it('gets value to single deep key', () => {
@@ -15,7 +14,7 @@ describe('#deepGet', () => {
     };
     const returnedValue = deepGet(initialObject, 'a');
 
-    expect(returnedValue).to.equal('test');
+    expect(returnedValue).toBe('test');
   });
 
   it('sets value to multi deep key', () => {
@@ -26,13 +25,13 @@ describe('#deepGet', () => {
     };
     const returnedValue = deepGet(initialObject, 'a.b');
 
-    expect(returnedValue).to.equal('test');
+    expect(returnedValue).toBe('test');
   });
 
   it('returns undefined for non existing value', () => {
     const initialObject = {};
     const returnedValue = deepGet(initialObject, 'a.b');
 
-    expect(returnedValue).to.be.undefined;
+    expect(returnedValue).toBeUndefined();
   });
 });
