+++ src/deepGet.ts
@@ -1,14 +1,14 @@
 export const deepGet = (object: object, key: string): any => {
-  const objectCopy = {...object};
-  if(!key) {
+  const objectCopy = { ...object };
+  if (!key) {
     return objectCopy;
   }
 
   const keys = key.split('.');
   let property = objectCopy;
-  for (let i = 0; i < keys.length; i++ ) {
+  for (let i = 0; i < keys.length; i++) {
     const currentKey = keys[i];
-    if(property) {
+    if (property) {
       property = property[currentKey];
     }
   }
