--- test/mocha.opts
@@ -1,2 +1,0 @@
-**/*.spec.ts
---compilers js:ts-node/register
