+++ tsconfig.json
@@ -1,25 +1,11 @@
 {
   "compilerOptions": {
-    "allowJs": true,
-    "module": "CommonJS",
+    "module": "es6",
     "moduleResolution": "Node",
-    "sourceMap": true,
-    "suppressImplicitAnyIndexErrors": true,
     "allowSyntheticDefaultImports": true,
     "target": "ES5"
   },
-  "files": [
-    "configurationLoader.ts"
-  ],
-  "typeRoots": [
-    "node_modules/@types"
-  ],
   "exclude": [
-    "node_modules",
-    "patterns",
-    "test",
-    "dist",
-    ".tmp",
-    "**/*.spec.ts"
+    "node_modules"
   ]
 }
